backend/requirements.txt
Django
psycopg2
graphene_django
flake8
factory_boy
django-cors-headers
django-filter
django-graphql-jwt
graphene_django_cud
faker
djangorestframework
django-model-utils
django-role-permissions